sob implementation and chaplain changes
A new job has been added. The Sister Elohiem. They are a Sister of Battle belonging to the Order of the Sacred Rose, marked as Elohiem. Half warrior, half preacher, they have willingly spilt blood in the Emperor's name and are one of His most devout and evangelical Sisters. They have been tasked with defending holy Ecclesiarchal relics and the Ecclesiarchy attached to them. Their main goal is to defend the chapel, the three holy relics inside, and the priest. Roleplay will be scrutinized. Requested by Nobel

The Abbot is now a Confessor-Millitant from the Missionarus Galaxia and has the privileges associated with it. Requested by Nick and backed by lore.