bolter pistol nerf, astartes authority clarification
The Bolter pistol has been nerfed, it used to do the same damage as a bolter rifle. it no longer does that.

Guardsmen report only to the Sergeant and Commissar and the Commissar reports only to the Rogue Trader, the Astartes Envoy no longer has official authority over the guard. requested by Nobel