Mortar Malfunctions
Mortars can malfunction now, either falling onto the ground and exploding or just becoming a dud, if it is a frag shell. Duds can be defused in two steps, and their chance to explode is very random, but they explode even if you step on, not when you step off.