Update to DUNGEON
This is a merge of my map from the other project.

Basically just adds the finishing touches to the dungeon, this was done literally in 2021 but never got patched in. So I just franksteined this bitch to the current map.

ADDITIONS

Sewer System.
Extended Slums & City Outskirts
Mines Area(Lots of lore & daemons to duke it out with.)

REMOVED
A few items.
Handful of daemons(Certain mobs would wander to the entrance of the dungeon. I don't like the idea of players dying before they even get inside the place, you should be able to step your foot in the door and THEN die.)

TIPS
If you are new to the dungeon, make sure you bring buddies. Understand the daemons move around a lot so they'll bump into you, if you want to roleplay or heal up. Find a place you can HIDE so you don't get your arms ripped off by the mobs patrolling the city. This place should scare you more then a Tau Strike Team, because you are more likely to survive The Tau then you are this.