Whip Fix
The whip does 5+30-40 damage per strike. It's a large downgrade but the whip has no cooldown on attacks. Meaning it hits as fast as you can click which is why people could decap the Astartes in half a second with this thing.

It's still OP as fuck 35-45 force combined with a high STR stat will still kill anyone who isn't wearing power armor.