Cargo Update
Increased prices of weaponry - especially Lasguns. One of the points in lore is Lasguns are expensive which is why PDF forces often get stubbers or autoguns instead of the legendary Las.

Armor is a little bit more expensive.

The commissar gear like the chainsword and bolt pistol are more expensive to discourage people buying them unless necessary.