Guns Update
Fixed fire delay on a number of weapons and also fixed that firing in automatic was actually slower for some guns then others.

Added the Inquisitor Bolt Pistol. Which is a faster firing and more accurate variant of the Garm Pattern.

Updated the Melta significantly to represent the heavy anti-tank weapon it is suppose to be. It has significant penalties for use in exchange for it's damage boost.