Bullets Update
Sidearms deal slightly more damage to encourage their use, too many characters walking around with LMG's and Lasrifles who should not be using them. This should hopefully encourage the use of non-two handed firearms more that aren't just a Bolt Pistol.

Also updated the Shotgun since everyone and their mother complains about how useless it is.

Rifles got a slight boost in damage, mainly the Bolt Action Rifles.

Ork Bullets deal a tiny bit more damage to be more of a threat to the Astartes. This is fair considering we nerfed the fuck out of their guns to miss 60 percent of the time.

Melta Gun now deals a lot of damage, it use to be absolute dogshit for damage. Literally worse then a Bolter in every way.