Computer Update
Lowers tithe cost and renames the Tithe Cogitator to Commerce Cogitator to be more lore friendly to a colony, they would not have dedicated tithe consoles on a colony that does not qualify for the Imperial Tithe yet.

It would be decades of work at the minimum before Eipharius could send anything of ANY value back to Holy Terra, for now they would be focusing on building the colony up. Instead the cogitator is dedicated to purchasing additional resources, in future you will exchange Thrones for Mercs/Supplies and maybe even more Guardsmen. Right now it's just a rename so we'll see.