Nid Talon
Uses the stunbaton code without any of the actual stunbaton stuff.

Effectively in game it's a claw weapon that stab you, causes bleeding but non fatal damage. It deals agony pain per strike but in total it's much less effective then using the regular nid talon weapon. It's not good at all, BUT it's an amazing tool for nids to capture people for converts.

In combat it's not a good idea to use, but if you catch someone alone you can spam them wit the Talon until they fall into agonycrit and then convert them without injuring your new cult member. Keep in mind how bad of an idea this is to do with victims who are shooting you.