Rebalance of bullets/beams
This was a longtime coming and anyone familiar with the old values will understand why we had to end this insanity.

ESPECIALLY with the overcharge values, nobody ever used lasguns on regular settings once we added overcharge and it's it was completely broken. 

Weapons are a lot more in line with a general balance that should encourage more versatile gunplay and be less punishing to people who are choosing to be melee characters or xenos.