Updated Pilgrim/Inquisition
Pilgrims should have less chance of critical failing in melee.

Interrogator slot reduced to 1. Inquisitor's should be more resourceful and stop relying on having stormtroopers do all the work for them. Also typically the rank of Interrogator is limited to one, as it's a grooming role to become an Inquisitor.

Scion is now structured into the Imperial Guard temporarily as an attachment to the RT's forces the info is as follows;
You are a Tempestus Scion attached to the retinue of the Lord Trader, your services ultimately being in the protection of imperial assets here on Eipharius on behalf of the Astra Militarum. It's barely been a week since leaving your Regiment and you are still adjusting to the new reality of your posting. Loyal to the Emperor above all else, Tempestus Scions provide the Imperial High Command with a core of flexible, nigh-incorruptible soldiers. Deployed sparingly, they can be depended upon to complete any mission, no matter how deadly. Racing into battle aboard armoured transports or plummeting down on Grav-Chutes, the Scions are the sharp edge of the Imperial Guard war machine.