You can now 'ghost'
Do not press the fucking ghost button if you are alive unless you want your brain to explode. This is an easy way out for people who get stuck in crit and is less immersive breaking also for people to just have a stroke then shoot themselves in the face to respawn. If you are at full health or are an Astartes, it will take more then one button press of Ghost to end your miserable life.