Guard Slots lowered
Since our rounds are heavily moderated, Guard Slots are lowered. This should allow staff to free up job slots if they want more Guard in the round or even Astartes without breaking the game. Since by default you could swarm the map with Guard Players.

This is intended to move in the direction of the hive city work which is to detract less from Warfare and more on the factional intrigue of 40k. Along with the commerce update which should focus rounds less on shooting things and more on telling stories.

Also removed the "You answer to the Commissar" text and reminded Commissar's not to be little powermad idiots.