Lateparty
Extended Lateparty Timer and added more timer options for admins. Basically gives us more control on when lateparty arrives and also set the default lateparty time to 60 minutes instead of 30. If admins are on they can set it to any time they want alternatively, will be especially important for when we switch to Messina Map.