Church + Magisterium updates
Improves a area in the Magisterium. Moves down the forensics area and adds a second stair.

Returns the gate to the church side-door.

Fixes the Arbite office in the magisterium to actually make sense, since they spawned backward and their locker was infront of the door.

Fixes accesses for Abbess door.