jukebox, alert levels, restores manor garden
so the jukebox had a bunch of copyrighted content on it so it has been replaced with old music that no longer has a copyright mostly from the 1910s-1950s.

Alert levels no longer mention the "Station" and have sufficiently draconian restrictions on citizens during a state of emergency for the grim, dark future of the 41st millenium

Makes novitiates labeled as "Sister" and the sister superior "Sister-superior" when they speak since it was causing confusion.

The copyrighted content has not been completely removed, I was just making sure it was not referenced in the code so it is easy to remove if you would like to.