map hotfix
fixes lighting in the garrison.
Fixes the garrison armory.

expands the ruined city to it's pre-edit state so it's harder to traverse, adds more obstacles, adds dirt instead of stone so you can make trenches on it but it mostly looks the same.

Adds a new tunnel for interrogators

Adds church books.

Gives left side clerk booth of the administratum access to both blast doors.